pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qiskit-addon-mpf"
version = "0.1.0"
readme = "README.md"
description = "Reducing the Trotter error of Hamiltonian dynamics with multi-product formulas"
license = {file = "LICENSE.txt"}
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]

requires-python = ">=3.9"

dependencies = [
    "numpy>=1.23",
    "cvxpy>=1.6",
    "qiskit>=1.2, <2",
]

[project.optional-dependencies]
quimb = [
    "quimb>=1.8.3, <2",
    "qiskit-quimb",
]
tenpy = [
    "physics-tenpy>=1.0.4, <2",
]
dynamic = [
    "qiskit-addon-mpf[quimb,tenpy]",
]
dev = [
    "qiskit-addon-mpf[test,doctest,nbtest,lint,docs]",
    "tox>=4.4.3",
]
basetest = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-subtests>=0.13",
    "qiskit-addon-mpf[dynamic]",
    "qiskit-addon-utils",
]
test = [
    "qiskit-addon-mpf[basetest]",
]
doctest = [
    "qiskit-addon-mpf[basetest,notebook-dependencies]",
    "pytest-doctestplus>=1.2.1",
]
nbtest = [
    "qiskit-addon-mpf[basetest]",
    "nbmake>=1.5.0",
]
style = [
    "ruff==0.7.1",
    "nbqa>=1.8.5",
    "typos>=1.20.0",
]
lint = [
    "qiskit-addon-mpf[style]",
    "mypy==1.13.0",
    "pylint==3.3.1",
    "reno>=4.1",
    "toml>=0.9.6",
]
notebook-dependencies = [
    "qiskit-addon-mpf",
    "rustworkx[graphviz]>=0.15",
    "matplotlib",
    "ipywidgets",
    "pylatexenc",
]
docs = [
    "qiskit-addon-mpf[doctest]",
    "qiskit-sphinx-theme~=2.0.0",
    "jupyter-sphinx",
    "sphinx-design",
    "sphinx-autodoc-typehints",
    "sphinx-copybutton",
    "sphinx_reredirects",
    "nbsphinx>=0.9.4",
    "reno>=4.1",
]

[tool.coverage.run]
branch = true
parallel = true

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.hatch.build.targets.wheel]
only-include = [
    "qiskit_addon_mpf",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
python_version = 3.9
show_error_codes = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pylint.main]
py-version = "3.9"

[tool.pylint."messages control"]
disable = ["all"]
enable = [
    "reimported",
    "no-else-raise",
    "redefined-argument-from-local",
    "redefined-builtin",
    "raise-missing-from",
    "cyclic-import",
    "unused-argument",
    "attribute-defined-outside-init",
    "no-else-return",
]

[tool.pytest.ini_options]
testpaths = ["./qiskit_addon_mpf/", "./test/"]

[tool.ruff]
line-length = 100
src = ["qiskit_addon_mpf", "test"]
target-version = "py39"

[tool.ruff.lint]
select = [
    "I",   # isort
    "E",   # pycodestyle
    "W",   # pycodestyle
    "D",   # pydocstyle
    "F",   # pyflakes
    "RUF", # ruff
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
]
ignore = [
    "E501", # line too long
]

[tool.ruff.lint.pylint]
max-args = 6

[tool.ruff.lint.extend-per-file-ignores]
"test/**.py" = [
    "D",  # pydocstyle
]
"docs/**/*" = [
    "E402", # module level import not at top of file
    "D",    # pydocstyle
]

[tool.ruff.lint.flake8-copyright]
notice-rgx = """
# This code is a Qiskit project.
#
# \\(C\\) Copyright IBM 2024\\.
#
# This code is licensed under the Apache License, Version 2\\.0\\. You may
# obtain a copy of this license in the LICENSE\\.txt file in the root directory
# of this source tree or at http\\:\\/\\/www\\.apache\\.org\\/licenses\\/LICENSE\\-2\\.0\\.
#
# Any modifications or derivative works of this code must retain this
# copyright notice, and modified files need to carry a notice indicating
# that they have been altered from the originals\\.
"""

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.typos.default.extend-words]
